보스 몬스터 hp ui 추가

+++ RogueLike/Assets/02.Scripts/Managers/GameManager.cs
@@ -97,7 +97,7 @@ public class GameManager : MonoBehaviour
     {
         DataManager.UpdatePlayerAttckSOData(atkSpeed, power, range);
     }
-    public void UpdatePlayerStatsDatas(int maxHealth,int health, int speed)
+    public void UpdatePlayerStatsDatas(int maxHealth,int health, float speed)
     {
         DataManager.UpdatePlayerStatsData(maxHealth ,health, speed);
     }
